Inicio de sesion - User , role, country class and repository
Creation of the classes of user, role and country. Also generate the repository of each one. At this point users can sign up when there is a country previously added at the DB. Controller, seeder and Enum also updated.

For now using a local DB. Must be changed for each developer or a cloud solution.